requirements.txt
flask
python-dotenv
requests
google-generativeai
youtube-transcript-api
langchain
pytube